+++ porky/src/porky_w.rs
@@ -7,10 +7,8 @@ use crate::hw_definition::description::{
     HardwareDescription, HardwareDetails, PinDescriptionSet, SsidSpec,
 };
 use crate::pin_descriptions::PIN_DESCRIPTIONS;
-#[cfg(feature = "wifi")]
 use crate::tcp::TCP_PORT;
 use core::str;
-#[cfg(feature = "wifi")]
 use cyw43_pio::PioSpi;
 use defmt::{error, info};
 use defmt_rtt as _;
@@ -32,14 +30,18 @@ use embassy_rp::usb::InterruptHandler as USBInterruptHandler;
 use embassy_rp::watchdog::Watchdog;
 use embassy_sync::blocking_mutex::raw::{NoopRawMutex, ThreadModeRawMutex};
 use embassy_sync::channel::Channel;
-//use embedded_io_async::Write;
 use heapless::Vec;
 use panic_probe as _;
 use static_cell::StaticCell;
 
 #[cfg(not(any(feature = "pico", feature = "pico_w")))]
 compile_error!(
-    "You must chose a feature from:[\"pico\", \"pico_w\"] to select a device to build for"
+    "You must chose a feature from [\"pico\", \"pico_w\"] to select a device to build for"
+);
+
+#[cfg(all(feature = "pico", feature = "pico_w"))]
+compile_error!(
+    "You must chose a just one of [\"pico\", \"pico_w\"] to select a device to build for"
 );
 
 /// The ssid config generated by build.rs in "$OUT_DIR/ssid.rs"
@@ -47,17 +49,12 @@ mod ssid {
     include!(concat!(env!("OUT_DIR"), "/ssid.rs"));
 }
 
-#[cfg(feature = "wifi")]
 /// Wi-Fi related functions
 mod wifi;
 
 #[cfg(feature = "usb")]
 mod usb;
 
-#[cfg(feature = "usb-raw")]
-mod usb_raw;
-
-#[cfg(feature = "wifi")]
 /// TCP related functions
 mod tcp;
 
@@ -94,10 +91,7 @@ pub static HARDWARE_EVENT_CHANNEL: Channel<ThreadModeRawMutex, HardwareConfigMes
 /// Create a [HardwareDescription] for this device with the provided serial number
 fn hardware_description(serial: &str) -> HardwareDescription {
     let details = HardwareDetails {
-        #[cfg(feature = "pico_w")]
         model: "Pi Pico W",
-        #[cfg(feature = "pico")]
-        model: "Pi Pico",
         hardware: "RP2040",
         revision: "",
         serial,
@@ -218,7 +212,7 @@ async fn main(spawner: Spawner) {
     static STATIC_BUF: StaticCell<[u8; 200]> = StaticCell::new();
     let static_buf = STATIC_BUF.init([0u8; 200]);
 
-    #[cfg(feature = "usb-raw")]
+    #[cfg(feature = "usb")]
     let watchdog = Watchdog::new(peripherals.WATCHDOG);
 
     // If we have a valid SsidSpec, then try and join that network using it
@@ -228,8 +222,8 @@ async fn main(spawner: Spawner) {
                 info!("Assigned IP: {}", ip);
 
                 let tcp = (ip.octets(), TCP_PORT);
-                #[cfg(feature = "usb-raw")]
-                usb_raw::start(spawner, driver, hw_desc, Some(tcp), db, watchdog).await;
+                #[cfg(feature = "usb")]
+                usb::start(spawner, driver, hw_desc, Some(tcp), db, watchdog).await;
 
                 let mut wifi_tx_buffer = [0; 4096];
                 let mut wifi_rx_buffer = [0; 4096];
@@ -274,14 +268,14 @@ async fn main(spawner: Spawner) {
             }
             Err(e) => {
                 error!("Could not join Wi-Fi network: {}, starting USB", e);
-                #[cfg(feature = "usb-raw")]
-                usb_raw::start(spawner, driver, hw_desc, None, db, watchdog).await;
+                #[cfg(feature = "usb")]
+                usb::start(spawner, driver, hw_desc, None, db, watchdog).await;
             }
         },
         None => {
             info!("No valid SsidSpec was found, starting USB alone");
-            #[cfg(feature = "usb-raw")]
-            usb_raw::start(spawner, driver, hw_desc, None, db, watchdog).await;
+            #[cfg(feature = "usb")]
+            usb::start(spawner, driver, hw_desc, None, db, watchdog).await;
         }
     }
 }
