+++ src/hw_definition/mod.rs
@@ -2,7 +2,7 @@
 pub mod config;
 pub mod description;
 pub mod pin_function;
-#[cfg(feature = "usb-raw")]
+#[cfg(feature = "usb")]
 pub mod usb_values;
 
 /// [BCMPinNumber] is used to refer to a GPIO pin by the Broadcom Chip Number
