+++ src/widgets/mod.rs
@@ -3,6 +3,6 @@ pub mod clicker;
 pub mod led;
 pub mod line;
 pub mod modal;
-#[cfg(any(feature = "iroh", feature = "tcp", feature = "usb-raw"))]
+#[cfg(any(feature = "iroh", feature = "tcp", feature = "usb"))]
 pub mod spinner;
 pub mod toast;
